--- hortum_mobile/lib/myapp.dart
@@ -1,22 +1,0 @@
-import 'package:adaptive_theme/adaptive_theme.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-
-import 'login_page.dart';
-
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return AdaptiveTheme(
-      // Adapta o tema
-      light: ThemeData.light(),
-      dark: ThemeData.dark(),
-      initial: AdaptiveThemeMode.system, //inicia o app com o tema do sistema
-      builder: (theme, darkTheme) => MaterialApp(
-        theme: theme,
-        darkTheme: darkTheme,
-        home: LoginPage(),
-      ),
-    );
-  }
-}
