+++ hortum_mobile/lib/view/features/login/login_page.dart
@@ -1,0 +1,245 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+import 'package:hortum_mobile/data/api_http.dart';
+
+class LoginPage extends StatefulWidget {
+  //LoginPage({Key key) : super(key: key);
+
+  @override
+  _LoginPageState createState() => _LoginPageState();
+}
+
+class _LoginPageState extends State<LoginPage> {
+  bool _isObscure = true;
+  TextEditingController emailController = new TextEditingController();
+  TextEditingController passwordController = new TextEditingController();
+  final _formKey = GlobalKey<FormState>();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Form(
+        key: _formKey,
+        child: ListView(
+          children: [
+            SizedBox(
+              height: MediaQuery.of(context).size.height / 7,
+            ),
+            logoImage(),
+            SizedBox(
+              height: MediaQuery.of(context).size.height / 22,
+            ),
+            emailTextForm(),
+            passwordTextForm(),
+            esqueciSenhaButton(),
+            entrarButton(),
+            naoTenhoContaButton(),
+          ],
+        ),
+      ),
+    );
+  }
+
+  Center logoImage() {
+    return Center(
+      child: Container(
+        height: MediaQuery.of(context).size.height / 4,
+        child: Align(
+          alignment: Alignment.center,
+          child: Image.asset("images/logo.png"),
+        ),
+      ),
+    );
+  }
+
+  Padding emailTextForm() {
+    return Padding(
+      padding: const EdgeInsets.fromLTRB(40, 8, 40, 8),
+      child: TextFormField(
+        //TextField email
+        controller: emailController,
+        validator: _validaLogin,
+        decoration: InputDecoration(
+          labelText: "Email",
+          labelStyle: TextStyle(
+            fontFamily: 'Roboto',
+            fontSize: 15,
+          ),
+          enabledBorder: UnderlineInputBorder(
+              borderSide: BorderSide(color: Color.fromARGB(0xFF, 244, 156, 0))),
+          focusedBorder: UnderlineInputBorder(
+              borderSide: BorderSide(color: Color.fromARGB(0xFF, 244, 156, 0))),
+          prefixIcon: Icon(
+            Icons.email,
+            color: Color.fromARGB(0xFF, 244, 156, 0), //Cor laranja
+          ),
+        ),
+      ),
+    );
+  }
+
+  Padding passwordTextForm() {
+    return Padding(
+      padding: const EdgeInsets.fromLTRB(40, 8, 40, 0),
+      child: TextFormField(
+        //TextFiled password
+        obscureText: _isObscure,
+        controller: passwordController,
+        validator: _validaPassword,
+        decoration: InputDecoration(
+          labelText: "Senha",
+          labelStyle: TextStyle(
+            fontFamily: 'Roboto',
+            fontSize: 15,
+          ),
+          enabledBorder: UnderlineInputBorder(
+              borderSide: BorderSide(color: Color.fromARGB(0xFF, 244, 156, 0))),
+          focusedBorder: UnderlineInputBorder(
+              borderSide: BorderSide(color: Color.fromARGB(0xFF, 244, 156, 0))),
+          prefixIcon: Icon(
+            Icons.lock,
+            color: Color.fromARGB(0xFF, 244, 156, 0),
+          ),
+          suffixIcon: IconButton(
+            icon: Icon(
+              _isObscure ? Icons.visibility_off : Icons.visibility,
+              color: Color.fromARGB(0xFF, 244, 156, 0),
+            ),
+            onPressed: () {
+              setState(() {
+                _isObscure = !_isObscure;
+              });
+            },
+          ),
+        ),
+      ),
+    );
+  }
+
+  String _validaLogin(String texto) {
+    if (texto.isEmpty) {
+      return "Digite o email";
+    }
+    return null;
+  }
+
+  String _validaPassword(String texto) {
+    if (texto.isEmpty) {
+      return "Digite a senha";
+    }
+    return null;
+  }
+
+  Padding esqueciSenhaButton() {
+    return Padding(
+      padding: const EdgeInsets.fromLTRB(180, 0, 0, 0),
+      child: TextButton(
+        //Botão esqueci a senha
+        onPressed: () {
+          print("esqueci minha senha");
+        },
+        child: Text(
+          "Esqueci minha senha",
+          style: TextStyle(
+            fontFamily: 'Roboto',
+            fontSize: 12,
+            color: Color.fromARGB(0xFF, 244, 156, 0),
+          ),
+        ),
+      ),
+    );
+  }
+
+  Padding entrarButton() {
+    return Padding(
+      padding: EdgeInsets.fromLTRB(85, 8, 85, 8),
+      child: ElevatedButton(
+        //Botão entrar
+        style: ElevatedButton.styleFrom(
+          primary: Color.fromARGB(0xFF, 244, 156, 0),
+          shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.all(Radius.circular(15))),
+        ),
+        onPressed: () {
+          onClickEntrar(context);
+        },
+        child: Text(
+          "ENTRAR",
+          style: TextStyle(
+            fontFamily: 'Roboto',
+            fontSize: 22,
+            package: 'fonts/Roboto/Roboto-Black.ttf',
+            color: Colors.white,
+          ),
+        ),
+      ),
+    );
+  }
+
+  Padding naoTenhoContaButton() {
+    return Padding(
+      padding: const EdgeInsets.fromLTRB(100, 6, 100, 8),
+      child: TextButton(
+        //Botão não tenho conta
+        onPressed: () {
+          print("Não tenho uma conta");
+        },
+        child: Text(
+          "Não tenho uma conta",
+          style: TextStyle(
+            fontFamily: 'Roboto',
+            fontSize: 12,
+            color: Color.fromARGB(0xFF, 244, 156, 0),
+          ),
+        ),
+      ),
+    );
+  }
+
+  onClickEntrar(BuildContext context) async {
+    bool formOK = _formKey.currentState.validate();
+    if (!formOK) return;
+
+    final email = emailController.text;
+    final password = passwordController.text;
+
+    print(email);
+    print(password);
+
+    var response = await LoginApi.login(email, password);
+
+    if (!response) {
+      showDialog(
+        context: context,
+        builder: (context) {
+          return AlertDialog(
+            title: Text("Erro!"),
+            content: Text(
+              "Email e/ou senha incorretos",
+              style: TextStyle(
+                fontFamily: 'Roboto',
+                fontSize: 16,
+              ),
+            ),
+            actions: [
+              // ignore: deprecated_member_use
+              FlatButton(
+                onPressed: () {
+                  Navigator.pop(context);
+                },
+                child: Text(
+                  "OK",
+                  style: TextStyle(
+                    fontFamily: 'Roboto',
+                    fontSize: 22,
+                    color: Color.fromARGB(0xFF, 244, 156, 0),
+                  ),
+                ),
+              )
+            ],
+          );
+        },
+      );
+    }
+  }
+}
